Fix(display): modify displaymodel
将屏幕获取缩放数据的方式更改，
如果GetScreenScaleFactors中包含该屏幕的缩放则使用，
否则使用 GetScaleFactor 中获取的数据

+++ src/frame/window/modules/display/displaywidget.cpp
@@ -175,14 +175,14 @@ void DisplayWidget::onMenuClicked(const QModelIndex &idx)
     }
 }
 
-int DisplayWidget::convertToSlider(const float value)
+int DisplayWidget::convertToSlider(const double value)
 {
     //remove base scale (100), then convert to 1-based value
     //with a stepping of 25
     return int(round(double(value * 100 - 100) / 25)) + 1;
 }
 
-float DisplayWidget::convertToScale(const int value)
+double DisplayWidget::convertToScale(const int value)
 {
-    return float(1.0 + (value - 1) * 0.25);
+    return 1.0 + (value - 1) * 0.25;
 }
